requirements.txt
pytest==7.4.2
dnspython==2.4.2
rich==13.6.0
pyfiglet==1.0.2
python-whois==0.8.0.